+++ code/qcommon/vm_local.h
@@ -127,7 +127,7 @@ struct vm_s {
     // DO NOT MOVE OR CHANGE THESE WITHOUT CHANGING THE VM_OFFSET_* DEFINES
     // USED BY THE ASM CODE
     int			programStack;		// the vm may be recursively entered
-    int			(*systemCall)( int *parms );
+    int			(*systemCall)( size_t *parms );
 
 	//------------------------------------
